% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flat_violin.R
\name{geom_flat_violin}
\alias{geom_flat_violin}
\title{geom_flat_violin}
\usage{
geom_flat_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "dodge",
  trim = TRUE,
  scale = "area",
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{aesthetic mapping}

\item{data}{the data to display in this layer}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{position adjustment}

\item{trim}{If ‘TRUE’ (default), trim the tails of the violins to the range of the data. If ‘FALSE’, don't trim the tails.}

\item{scale}{if "area" (default), all violins have the same area (before trimming the tails). If "count", areas are scaled proportionally to the number of observations. If "width", all violins have the same maximum width.}

\item{show.legend}{whether show the legend of this layer}

\item{inherit.aes}{whether inherit aesthetic mapping from `ggplot`}

\item{...}{additional parameters}
}
\value{
ggplot layer
}
\description{
flat violin geom
}
\examples{
library(ggplot2)
ggplot(diamonds, aes(cut, carat)) +
   geom_flat_violin() +
   coord_flip()
}
\author{
David Robinson
}
