% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ord_ellipse.R
\name{geom_ord_ellipse}
\alias{geom_ord_ellipse}
\title{add confidence ellipse to ordinary plot}
\usage{
geom_ord_ellipse(mapping = NULL, ellipse_pro = 0.97, fill = NA, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{ellipse_pro}{confidence value for the ellipse}

\item{fill}{color to fill the ellipse, NA by default}

\item{...}{additional parameters}
}
\value{
ggplot layer
}
\description{
add confidence ellipse to ordinary plot produced by ggord
}
\author{
Guangchuang Yu
}
