% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_triangle.R
\name{geom_triangle}
\alias{geom_triangle}
\title{geom_triangle}
\usage{
geom_triangle(mapping = NULL, data = NULL, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{...}{additional parameters}
}
\value{
ggplot2 layer
}
\description{
ggplot2 layer of triangle
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, disp)) + geom_triangle()
}
\author{
Shipeng Guo
}
