% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ord_ellipse.R
\name{geom_ord_ellipse}
\alias{geom_ord_ellipse}
\title{add confidence ellipse to ordinary plot}
\usage{
geom_ord_ellipse(mapping = NULL, ellipse_pro = 0.97, fill = NA, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{ellipse_pro}{confidence value for the ellipse}

\item{fill}{color to fill the ellipse, NA by default}

\item{...}{additional parameters}
}
\value{
ggplot layer
}
\description{
add confidence ellipse to ordinary plot produced by ggord
}
\examples{
\dontrun{
library(MASS)
ord <- lda(Species ~ ., iris, prior = rep(1, 3)/3)
## devtools::install_github('fawda123/ggord')
library(ggord)
p <- ggord(ord, iris$Species)
p + geom_ord_ellipse(ellipse_pro = .96, color='firebrick', size=1, lty=3) +
geom_ord_ellipse(ellipse_pro = .99, lty=2)
}
}
\author{
Guangchuang Yu
}
