% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_segment_c.R
\name{geom_segment_c}
\alias{geom_segment_c}
\title{geom_segment_c}
\usage{
geom_segment_c(mapping = NULL, data = NULL, position = "identity",
  lineend = "butt", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, arrow = NULL, arrow.fill = NULL, ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{position}{position}

\item{lineend}{lineend}

\item{na.rm}{logical}

\item{show.legend}{logical}

\item{inherit.aes}{logical}

\item{arrow}{specification for arrow heads, as created by arrow().}

\item{arrow.fill}{fill color to usse for the arrow head (if closed). `NULL` means use `colour` aesthetic.}

\item{...}{additional parameter}
}
\value{
add segment layer
}
\description{
geom_segment_c supports coloring segment with continuous colors
}
\seealso{
\link[ggplot2]{geom_segment}
}
\author{
Guangchuang Yu
}
